[FEAT] : 공통 로직 추가 KAN-127
