+++ apps/main-web/package.json
@@ -16,6 +16,7 @@
     "lucide-react": "^0.454.0",
     "next": "^14.2.6",
     "next-auth": "^4.24.8",
+    "next-s3-upload": "^0.3.4",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-hook-form": "^7.53.1",
