+++ apps/main-web/src/actions/feed/FeedCard.ts
@@ -8,6 +8,7 @@ export const getFeedCardList = async () => {
 };
 
 export const uploadToS3 = async (file: File) => {
+  console.log(File);
   const session = await getServerSession(options);
   const user_uuid = session?.user?.uuid;
   console.log("파일 업로드할껀데  uuid가?", user_uuid);
