+++ apps/main-web/src/app/(onlybottom)/addfeed/page.tsx
@@ -2,7 +2,7 @@ import AddFeed from "../../../components/feedcard/AddFeed";
 
 export default function page() {
   return (
-    <div className="mb-14">
+    <div className="mb-5">
       <AddFeed></AddFeed>
     </div>
   );
