+++ apps/main-web/src/app/api/s3-upload/route.ts
@@ -1,0 +1,1 @@
+export { POST } from "next-s3-upload/route";
